src/cycleroom/requirements.txt

fastapi
uvicorn
httpx
pydantic
influxdb-client
asyncpg
pytest
flake8
black
load_dotenv